Move tooltip generation out of BoardView
